ibatis存在SQL缓存，所以自定义字段的时候需要开启 remapResults="true" 存在一定性能问题  <select id="get${tbb.className}ListWithPageFields" parameterClass="${tbb.className}Query" remapResults="true" resultClass="${tbb.className}">

+++ src/main/java/com/raycloud/util/daogen/SqlType2Feild.java
@@ -22,8 +22,7 @@ public class SqlType2Feild {
 		sqltype2Feild.put(Types.VARCHAR, "String");
 		sqltype2Feild.put(Types.DATE, "Date");	//java.util.Date
 		//mysql
-		sqltype2Feild.put(Types.BIT, "Byte");
-//		sqltype2Feild.put(Types.TINYINT, "Byte");
+		sqltype2Feild.put(Types.BIT, "Long");
         sqltype2Feild.put(Types.TINYINT, "Long");
 		sqltype2Feild.put(Types.SMALLINT, "Integer");
 		sqltype2Feild.put(Types.INTEGER, "Integer");
